backend/requirements.txt
# FastAPI & server
fastapi==0.110.0
uvicorn[standard]==0.29.0

# Database
SQLAlchemy==2.0.28
psycopg[binary,pool]==3.1.18
sqlalchemy-utils==0.41.2
pgvector==0.2.5
psycopg2
python-jose[cryptography]
passlib[bcript]
python-multipart

# Haystack 2.x + integrations
haystack-ai[pgvector]==2.2.2
haystack-ai[integrations]

# OpenAI API
openai==1.12.0

# --- Optional / Useful ---
pydantic>=2.7.0            # Data validation

# Hugging Face + local embeddings
transformers==4.37.2
torch>=2.1.0
sentence-transformers==2.2.2